consolidating examples
