+++ examples/animation/image_slices_viewer.py
