+++ examples/images_contours_and_fields/image_masked.py
