+++ examples/images_contours_and_fields/image_nonuniform.py
