+++ examples/pylab_examples/fancybox_demo.py
@@ -3,11 +3,37 @@
 Fancybox Demo
 =============
 
+Plotting fancy boxes with Matplotlib.
+
+The following examples show how to plot boxes with different
+visual properties.
 """
 import matplotlib.pyplot as plt
 import matplotlib.transforms as mtransforms
+import matplotlib.patches as mpatch
 from matplotlib.patches import FancyBboxPatch
 
+###############################################################################
+# First we'll show some sample boxes with fancybox.
+
+styles = mpatch.BoxStyle.get_styles()
+spacing = 1.2
+
+figheight = (spacing * len(styles) + .5)
+fig1 = plt.figure(1, (4/1.5, figheight/1.5))
+fontsize = 0.3 * 72
+
+for i, stylename in enumerate(sorted(styles)):
+    fig1.text(0.5, (spacing * (len(styles) - i) - 0.5) / figheight, stylename,
+              ha="center",
+              size=fontsize,
+              transform=fig1.transFigure,
+              bbox=dict(boxstyle=stylename, fc="w", ec="k"))
+
+plt.show()
+
+###############################################################################
+# Next we'll show off multiple fancy boxes at once.
 
 # Bbox object around which the fancy box will be drawn.
 bb = mtransforms.Bbox([[0.3, 0.4], [0.7, 0.6]])
