--- examples/pylab_examples/fancybox_demo2.py
@@ -1,24 +1,0 @@
-"""
-==============
-Fancybox Demo2
-==============
-
-"""
-import matplotlib.patches as mpatch
-import matplotlib.pyplot as plt
-
-styles = mpatch.BoxStyle.get_styles()
-spacing = 1.2
-
-figheight = (spacing * len(styles) + .5)
-fig1 = plt.figure(1, (4/1.5, figheight/1.5))
-fontsize = 0.3 * 72
-
-for i, stylename in enumerate(sorted(styles)):
-    fig1.text(0.5, (spacing * (len(styles) - i) - 0.5) / figheight, stylename,
-              ha="center",
-              size=fontsize,
-              transform=fig1.transFigure,
-              bbox=dict(boxstyle=stylename, fc="w", ec="k"))
-
-plt.show()
