--- examples/pylab_examples/hexbin_demo.py
@@ -1,42 +1,0 @@
-"""
-===========
-Hexbin Demo
-===========
-
-hexbin is an axes method or pyplot function that is essentially
-a pcolor of a 2-D histogram with hexagonal cells.  It can be
-much more informative than a scatter plot; in the first subplot
-below, try substituting 'scatter' for 'hexbin'.
-"""
-
-import numpy as np
-import matplotlib.pyplot as plt
-
-# Fixing random state for reproducibility
-np.random.seed(19680801)
-
-n = 100000
-x = np.random.standard_normal(n)
-y = 2.0 + 3.0 * x + 4.0 * np.random.standard_normal(n)
-xmin = x.min()
-xmax = x.max()
-ymin = y.min()
-ymax = y.max()
-
-fig, axs = plt.subplots(ncols=2, sharey=True, figsize=(7, 4))
-fig.subplots_adjust(hspace=0.5, left=0.07, right=0.93)
-ax = axs[0]
-hb = ax.hexbin(x, y, gridsize=50, cmap='inferno')
-ax.axis([xmin, xmax, ymin, ymax])
-ax.set_title("Hexagon binning")
-cb = fig.colorbar(hb, ax=ax)
-cb.set_label('counts')
-
-ax = axs[1]
-hb = ax.hexbin(x, y, gridsize=50, bins='log', cmap='inferno')
-ax.axis([xmin, xmax, ymin, ymax])
-ax.set_title("With a log color scale")
-cb = fig.colorbar(hb, ax=ax)
-cb.set_label('log10(N)')
-
-plt.show()
