--- examples/pylab_examples/legend_demo2.py
@@ -1,26 +1,0 @@
-"""
-============
-Legend Demo2
-============
-
-Make a legend for specific lines.
-"""
-import matplotlib.pyplot as plt
-import numpy as np
-
-
-t1 = np.arange(0.0, 2.0, 0.1)
-t2 = np.arange(0.0, 2.0, 0.01)
-
-# note that plot returns a list of lines.  The "l1, = plot" usage
-# extracts the first element of the list into l1 using tuple
-# unpacking.  So l1 is a Line2D instance, not a sequence of lines
-l1, = plt.plot(t2, np.exp(-t2))
-l2, l3 = plt.plot(t2, np.sin(2 * np.pi * t2), '--o', t1, np.log(1 + t1), '.')
-l4, = plt.plot(t2, np.exp(-t2) * np.sin(2 * np.pi * t2), 's-.')
-
-plt.legend((l2, l4), ('oscillatory', 'damped'), loc='upper right', shadow=True)
-plt.xlabel('time')
-plt.ylabel('volts')
-plt.title('Damped oscillation')
-plt.show()
