--- examples/pylab_examples/legend_demo3.py
@@ -1,28 +1,0 @@
-"""
-============
-Legend Demo3
-============
-
-"""
-import matplotlib.pyplot as plt
-import numpy as np
-
-x = np.linspace(0, 1)
-
-# Plot the lines y=x**n for n=1..4.
-ax = plt.subplot(2, 1, 1)
-for n in range(1, 5):
-    plt.plot(x, x**n, label="n={0}".format(n))
-plt.legend(loc="upper left", bbox_to_anchor=[0, 1],
-           ncol=2, shadow=True, title="Legend", fancybox=True)
-ax.get_legend().get_title().set_color("red")
-
-# Demonstrate some more complex labels.
-ax = plt.subplot(2, 1, 2)
-plt.plot(x, x**2, label="multi\nline")
-half_pi = np.linspace(0, np.pi / 2)
-plt.plot(np.sin(half_pi), np.cos(half_pi), label=r"$\frac{1}{2}\pi$")
-plt.plot(x, 2**(x**2), label="$2^{x^2}$")
-plt.legend(shadow=True, fancybox=True)
-
-plt.show()
