--- examples/pylab_examples/legend_demo4.py
@@ -1,28 +1,0 @@
-"""
-============
-Legend Demo4
-============
-
-"""
-import matplotlib.pyplot as plt
-
-fig, axes = plt.subplots(3, 1)
-top_ax, middle_ax, bottom_ax = axes
-
-top_ax.bar([0, 1, 2], [0.2, 0.3, 0.1], width=0.4, label="Bar 1",
-           align="center")
-top_ax.bar([0.5, 1.5, 2.5], [0.3, 0.2, 0.2], color="red", width=0.4,
-           label="Bar 2", align="center")
-top_ax.legend()
-
-middle_ax.errorbar([0, 1, 2], [2, 3, 1], xerr=0.4, fmt="s", label="test 1")
-middle_ax.errorbar([0, 1, 2], [3, 2, 4], yerr=0.3, fmt="o", label="test 2")
-middle_ax.errorbar([0, 1, 2], [1, 1, 3], xerr=0.4, yerr=0.3, fmt="^",
-                   label="test 3")
-middle_ax.legend()
-
-bottom_ax.stem([0.3, 1.5, 2.7], [1, 3.6, 2.7], label="stem test")
-bottom_ax.legend()
-
-plt.subplots_adjust(hspace=0.7)
-plt.show()
